Extract I/O into a separate module
