+++ .gitignore
@@ -214,3 +214,4 @@ requirements.txt
 
 # Ignore raw data files
 src/data/**
+src/tmpvenv/**
