+++ src/analyzer/main.py
@@ -21,7 +21,7 @@ def main():
     else:
         logger.debug("File loaded: %s", file_path)
 
-    finder = package_metadata_fetcher.PyMetadataBuilder(file_path)
+    finder = package_metadata_fetcher.build_package_metadata(file_path)
     for pkg in finder:
         print(f"{pkg.package} | {pkg.license_type} | {pkg.link}")
